playhead-tick major bugfix: do not allow ticks to jump over a beat
playhead-tick fires when the incoming position occurs after a user-
specified threshold within the beat. But it didn't "freeze" the
position variable at this moment. So it would delay the outgoing
tick to occur on the beat, but the beat value to be output could
be later than expected. Sometimes this would cause a beat to be skipped
and the next beat repeated, e.g. 100, 101, 103, 103, 104.

Fixed.